Fixed reset

+++ reset.exclude.zsh
@@ -12,6 +12,7 @@ check_args "$@"
 confirm_args "$@"
 rm_dotfiles "$@"
 
+confirm "Delete zgen, oh-my-zsh and .vim? [Y/n]" || exit 0
 [ -d "$dotfiles/.zgen" ] && rm -rf $dotfiles/.zgen
 [ -d "$dotfiles/.oh-my-zsh" ] && rm -rf $dotfiles/.oh-my-zsh
 [ -d "$dotfiles/.vim" ] && rm -rf $dotfiles/.vim
